requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
httpx==0.28.1
openai==1.58.1
mcp>=1.0.0
python-dotenv==1.0.1
python-ulid==3.0.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
jinja2==3.1.4
pgvector==0.3.6
pytest==8.3.4
pytest-asyncio==0.24.0
pyyaml==6.0.2
